removed alignment and added feederShoot
